refined clean_date
